feat: implement DeleteCourseUseCase

+++ src/main/java/br/com/rocketseat/challenger01/course/modules/useCases/DeleteCourseUseCase.java
@@ -19,6 +19,12 @@ public class DeleteCourseUseCase {
       throw new Exception("Resource not found");
     }
 
+    var course = this.courseRepository.findById(id);
+
+    if(course.isEmpty()){
+      throw new Exception("Resource not found");
+    }
+
     this.courseRepository.deleteById(id);
   }
 }
